requirements.txt
PyPDF2>=3.0.0
pandas>=1.5.0
openpyxl>=3.0.0
python-dateutil>=2.8.0
regex>=2022.0.0
cryptography>=3.4.0 